Refactor gas station example and update environment configuration
- Changed SKIP_AUTHORIZATION flag in `.env.local.example` from "true" to "false" to enable authorization.
- Removed obsolete `combined_abi.ts` and `gassy-abi` files, consolidating ABI definitions into `gas-station.ts`.
- Updated imports in `transferETHDelegated.ts` and `transferUSDCDelegated.ts` to use the new `gas-station` ABI.
- Streamlined environment variable handling for dual contract approach in the gas station example.
