+++ examples/tk-gas-station/.env.local.example
@@ -11,7 +11,7 @@ BASE_RPC_URL="<>"
 DELEGATE_CONTRACT="<>"
 EXECUTION_CONTRACT="<>"
 # skips EIP 7702 Authorization
-SKIP_AUTHORIZATION="true"
+SKIP_AUTHORIZATION="false"
