--- examples/tk-gas-station/abi/combined_abi.ts
@@ -1,360 +1,0 @@
-export const gasStationAbi = [
-  { inputs: [], stateMutability: "nonpayable", type: "constructor" },
-  { inputs: [], name: "BatchSizeExceeded", type: "error" },
-  { inputs: [], name: "DeadlineExceeded", type: "error" },
-  { inputs: [], name: "ExecutionFailed", type: "error" },
-  { inputs: [], name: "InvalidCounter", type: "error" },
-  { inputs: [], name: "InvalidNonce", type: "error" },
-  { inputs: [], name: "InvalidOutputContract", type: "error" },
-  { inputs: [], name: "NotSelf", type: "error" },
-  {
-    inputs: [],
-    name: "MAX_BATCH_SIZE",
-    outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "burnNonce",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "uint128", name: "_nonce", type: "uint128" },
-      { internalType: "bytes", name: "_signature", type: "bytes" },
-    ],
-    name: "burnNonce",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "uint128", name: "_counter", type: "uint128" },
-      { internalType: "address", name: "_sender", type: "address" },
-      { internalType: "bytes", name: "_signature", type: "bytes" },
-    ],
-    name: "burnTimeboxedCounter",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "burnTimeboxedCounter",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "eip712Domain",
-    outputs: [
-      { internalType: "bytes1", name: "fields", type: "bytes1" },
-      { internalType: "string", name: "name", type: "string" },
-      { internalType: "string", name: "version", type: "string" },
-      { internalType: "uint256", name: "chainId", type: "uint256" },
-      { internalType: "address", name: "verifyingContract", type: "address" },
-      { internalType: "bytes32", name: "salt", type: "bytes32" },
-      { internalType: "uint256[]", name: "extensions", type: "uint256[]" },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "uint128", name: "_nonce", type: "uint128" },
-      { internalType: "address", name: "_outputContract", type: "address" },
-      { internalType: "uint256", name: "_ethAmount", type: "uint256" },
-      { internalType: "bytes", name: "_arguments", type: "bytes" },
-      { internalType: "bytes", name: "_signature", type: "bytes" },
-    ],
-    name: "execute",
-    outputs: [
-      { internalType: "bool", name: "", type: "bool" },
-      { internalType: "bytes", name: "", type: "bytes" },
-    ],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "uint128", name: "_nonce", type: "uint128" },
-      { internalType: "address", name: "_outputContract", type: "address" },
-      { internalType: "bytes", name: "_arguments", type: "bytes" },
-      { internalType: "bytes", name: "_signature", type: "bytes" },
-    ],
-    name: "execute",
-    outputs: [
-      { internalType: "bool", name: "", type: "bool" },
-      { internalType: "bytes", name: "", type: "bytes" },
-    ],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "uint128", name: "_nonce", type: "uint128" },
-      {
-        components: [
-          { internalType: "address", name: "outputContract", type: "address" },
-          { internalType: "uint256", name: "ethAmount", type: "uint256" },
-          { internalType: "bytes", name: "arguments", type: "bytes" },
-        ],
-        internalType: "struct IBatchExecution.Execution[]",
-        name: "_executions",
-        type: "tuple[]",
-      },
-      { internalType: "bytes", name: "_signature", type: "bytes" },
-    ],
-    name: "executeBatch",
-    outputs: [
-      { internalType: "bool", name: "", type: "bool" },
-      { internalType: "bytes[]", name: "", type: "bytes[]" },
-    ],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "uint128", name: "_counter", type: "uint128" },
-      { internalType: "uint128", name: "_deadline", type: "uint128" },
-      { internalType: "address", name: "_outputContract", type: "address" },
-      {
-        components: [
-          { internalType: "address", name: "outputContract", type: "address" },
-          { internalType: "uint256", name: "ethAmount", type: "uint256" },
-          { internalType: "bytes", name: "arguments", type: "bytes" },
-        ],
-        internalType: "struct IBatchExecution.Execution[]",
-        name: "_executions",
-        type: "tuple[]",
-      },
-      { internalType: "bytes", name: "_signature", type: "bytes" },
-    ],
-    name: "executeBatchTimeboxed",
-    outputs: [
-      { internalType: "bool", name: "", type: "bool" },
-      { internalType: "bytes[]", name: "", type: "bytes[]" },
-    ],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "uint128", name: "_counter", type: "uint128" },
-      { internalType: "uint128", name: "_deadline", type: "uint128" },
-      {
-        components: [
-          { internalType: "address", name: "outputContract", type: "address" },
-          { internalType: "uint256", name: "ethAmount", type: "uint256" },
-          { internalType: "bytes", name: "arguments", type: "bytes" },
-        ],
-        internalType: "struct IBatchExecution.Execution[]",
-        name: "_executions",
-        type: "tuple[]",
-      },
-      { internalType: "bytes", name: "_signature", type: "bytes" },
-    ],
-    name: "executeBatchTimeboxedArbitrary",
-    outputs: [
-      { internalType: "bool", name: "", type: "bool" },
-      { internalType: "bytes[]", name: "", type: "bytes[]" },
-    ],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "uint128", name: "_counter", type: "uint128" },
-      { internalType: "uint128", name: "_deadline", type: "uint128" },
-      { internalType: "address", name: "_outputContract", type: "address" },
-      { internalType: "bytes", name: "_arguments", type: "bytes" },
-      { internalType: "bytes", name: "_signature", type: "bytes" },
-    ],
-    name: "executeTimeboxed",
-    outputs: [
-      { internalType: "bool", name: "", type: "bool" },
-      { internalType: "bytes", name: "", type: "bytes" },
-    ],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "uint128", name: "_counter", type: "uint128" },
-      { internalType: "uint128", name: "_deadline", type: "uint128" },
-      { internalType: "address", name: "_outputContract", type: "address" },
-      { internalType: "uint256", name: "_ethAmount", type: "uint256" },
-      { internalType: "bytes", name: "_arguments", type: "bytes" },
-      { internalType: "bytes", name: "_signature", type: "bytes" },
-    ],
-    name: "executeTimeboxed",
-    outputs: [
-      { internalType: "bool", name: "", type: "bool" },
-      { internalType: "bytes", name: "", type: "bytes" },
-    ],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "uint128", name: "_counter", type: "uint128" },
-      { internalType: "uint128", name: "_deadline", type: "uint128" },
-      { internalType: "address", name: "_outputContract", type: "address" },
-      { internalType: "uint256", name: "_ethAmount", type: "uint256" },
-      { internalType: "bytes", name: "_arguments", type: "bytes" },
-      { internalType: "bytes", name: "_signature", type: "bytes" },
-    ],
-    name: "executeTimeboxedArbitrary",
-    outputs: [
-      { internalType: "bool", name: "", type: "bool" },
-      { internalType: "bytes", name: "", type: "bytes" },
-    ],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "uint128", name: "_counter", type: "uint128" },
-      { internalType: "uint128", name: "_deadline", type: "uint128" },
-      { internalType: "address", name: "_outputContract", type: "address" },
-      { internalType: "bytes", name: "_arguments", type: "bytes" },
-      { internalType: "bytes", name: "_signature", type: "bytes" },
-    ],
-    name: "executeTimeboxedArbitrary",
-    outputs: [
-      { internalType: "bool", name: "", type: "bool" },
-      { internalType: "bytes", name: "", type: "bytes" },
-    ],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "uint128", name: "_counter", type: "uint128" },
-      { internalType: "uint128", name: "_deadline", type: "uint128" },
-      { internalType: "address", name: "_sender", type: "address" },
-    ],
-    name: "hashArbitraryTimeboxedExecution",
-    outputs: [{ internalType: "bytes32", name: "", type: "bytes32" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "uint128", name: "_nonce", type: "uint128" },
-      {
-        components: [
-          { internalType: "address", name: "outputContract", type: "address" },
-          { internalType: "uint256", name: "ethAmount", type: "uint256" },
-          { internalType: "bytes", name: "arguments", type: "bytes" },
-        ],
-        internalType: "struct IBatchExecution.Execution[]",
-        name: "_executions",
-        type: "tuple[]",
-      },
-    ],
-    name: "hashBatchExecution",
-    outputs: [{ internalType: "bytes32", name: "", type: "bytes32" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [{ internalType: "uint128", name: "_nonce", type: "uint128" }],
-    name: "hashBurnNonce",
-    outputs: [{ internalType: "bytes32", name: "", type: "bytes32" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "uint128", name: "_counter", type: "uint128" },
-      { internalType: "address", name: "_sender", type: "address" },
-    ],
-    name: "hashBurnTimeboxedCounter",
-    outputs: [{ internalType: "bytes32", name: "", type: "bytes32" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "uint128", name: "_nonce", type: "uint128" },
-      { internalType: "address", name: "_outputContract", type: "address" },
-      { internalType: "uint256", name: "_ethAmount", type: "uint256" },
-      { internalType: "bytes", name: "_arguments", type: "bytes" },
-    ],
-    name: "hashExecution",
-    outputs: [{ internalType: "bytes32", name: "", type: "bytes32" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "uint128", name: "_counter", type: "uint128" },
-      { internalType: "uint128", name: "_deadline", type: "uint128" },
-      { internalType: "address", name: "_sender", type: "address" },
-      { internalType: "address", name: "_outputContract", type: "address" },
-    ],
-    name: "hashTimeboxedExecution",
-    outputs: [{ internalType: "bytes32", name: "", type: "bytes32" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "nonce",
-    outputs: [{ internalType: "uint128", name: "", type: "uint128" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "address", name: "", type: "address" },
-      { internalType: "address", name: "", type: "address" },
-      { internalType: "uint256[]", name: "", type: "uint256[]" },
-      { internalType: "uint256[]", name: "", type: "uint256[]" },
-      { internalType: "bytes", name: "", type: "bytes" },
-    ],
-    name: "onERC1155BatchReceived",
-    outputs: [{ internalType: "bytes4", name: "", type: "bytes4" }],
-    stateMutability: "pure",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "address", name: "", type: "address" },
-      { internalType: "address", name: "", type: "address" },
-      { internalType: "uint256", name: "", type: "uint256" },
-      { internalType: "uint256", name: "", type: "uint256" },
-      { internalType: "bytes", name: "", type: "bytes" },
-    ],
-    name: "onERC1155Received",
-    outputs: [{ internalType: "bytes4", name: "", type: "bytes4" }],
-    stateMutability: "pure",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "address", name: "", type: "address" },
-      { internalType: "address", name: "", type: "address" },
-      { internalType: "uint256", name: "", type: "uint256" },
-      { internalType: "bytes", name: "", type: "bytes" },
-    ],
-    name: "onERC721Received",
-    outputs: [{ internalType: "bytes4", name: "", type: "bytes4" }],
-    stateMutability: "pure",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "timeboxedCounter",
-    outputs: [{ internalType: "uint128", name: "", type: "uint128" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  { stateMutability: "payable", type: "receive" },
-] as const;
