+++ examples/tk-gas-station/abi/gas-station.ts
